LayoutTests/java/lc3/ConvertString/string-007-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion: string

PASS TEST_CLASS["ambiguous(boolean)"](string) + '' threw exception TypeError: Result of expression 'TEST_CLASS["ambiguous(boolean)"]' [undefined] is not a function..
PASS successfullyParsed is true

TEST COMPLETE
 
